refactor: improve route handling
