+++ Dockerfile
@@ -7,7 +7,6 @@ COPY package*.json ./
 COPY . .
 
 RUN npm install
-RUN npm build
 
 EXPOSE 8082
