+++ src/HackathonAPI.ts
@@ -2,7 +2,7 @@ import 'reflect-metadata';
 import express from 'express';
 import { UserController, TeamController, DiscordController } from './controllers';
 import RouteHandler, { ExpressHandler } from './RouteHandler';
-import { UsersRoute, UserRoute, TeamsRoute, TeamRoute, DiscordRoute } from './routes';
+import * as routes from './routes';
 import { createConnection, ConnectionOptions, Connection } from 'typeorm';
 import { User } from './entities/User';
 import { Team } from './entities/Team';
@@ -85,11 +85,9 @@ export default class HackathonAPI {
 	private configureExpress() {
 		return new Promise((resolve, reject) => {
 			this.router.use(express.json());
-			this.addRoute(new UsersRoute(this));
-			this.addRoute(new UserRoute(this));
-			this.addRoute(new TeamsRoute(this));
-			this.addRoute(new TeamRoute(this));
-			this.addRoute(new DiscordRoute(this));
+			for (const Route of Object.values(routes)) {
+				this.addRoute(new Route(this));
+			}
 			this.express.listen(this.options.api.port, resolve);
 			this.express.on('error', err => {
 				reject(err);
@@ -103,6 +101,7 @@ export default class HackathonAPI {
 		const path = handler.getRoute();
 		for (const verb of VERBS) {
 			const fn = handler[verb];
+			this.options.loggers.api.info(`Registered ${verb.toUpperCase()} ${handler.getRoute()}`);
 			if (fn) this.router[verb](path, this.wrapHandler(fn.bind(handler)));
 		}
 	}
