+++ src/routes/teams/index.ts
@@ -1,0 +1,2 @@
+export * from './team';
+export * from './teams';
