+++ src/routes/users/index.ts
@@ -1,0 +1,2 @@
+export * from './user';
+export * from './users';
