+++ src/routes/users/user.ts
@@ -1,5 +1,5 @@
-import RouteHandler from '../RouteHandler';
-import HackathonAPI from '../HackathonAPI';
+import RouteHandler from '../../RouteHandler';
+import HackathonAPI from '../../HackathonAPI';
 import { Request, Response } from 'express';
 
 export class UserRoute implements RouteHandler {
